backend/my_project/requirements.txt
# Django and REST Framework
Django==5.2.4
djangorestframework==3.15.2

# Swagger/OpenAPI Documentation
drf-yasg==1.21.7

# CORS handling
django-cors-headers==4.3.1

# OCR and Image Processing - EasyOCR
easyocr>=1.7.0
Pillow>=10.0.0
numpy>=1.26.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
pdf2image>=1.17.0  # For PDF to image conversion (requires Poppler)

# Development and testing dependencies
requests>=2.31.0

# QR Code generation
qrcode[pil]>=7.4.2

# PostgreSQL database adapter
psycopg2>=2.9.7

# HTML parsing (for HTML -> text indexing)
beautifulsoup4>=4.12.0

# Optional: Additional image processing
# scikit-image>=0.21.0

# Development and testing (optional)
# pytest>=7.4.0
# pytest-django>=4.5.0 